LayoutTests/fast/events/touch/frame-hover-update-expected.txt

PASS [object HTMLDocument] is non-null.
PASS [object HTMLDocument] is non-null.
PASS hoveredTop.length is 1
PASS hoveredIframe1.length is 1
PASS hoveredIframe2.length is 0
PASS hoveredTop.length is 1
PASS hoveredIframe1.length is 0
PASS hoveredIframe2.length is 1
PASS hoveredTop.length is 0
PASS hoveredIframe1.length is 0
PASS hoveredIframe2.length is 0
PASS hoveredTop.length is 1
PASS hoveredIframe1.length is 0
PASS hoveredIframe2.length is 0
PASS hoveredTop.length is 1
PASS hoveredIframe1.length is 0
PASS hoveredIframe2.length is 0
PASS successfullyParsed is true

TEST COMPLETE
